Refactored unify ranges into separate file
