+++ src/util/unifyRanges.ts
@@ -1,4 +1,7 @@
-import { Range } from "vscode";
+import { Range, Selection } from "vscode";
+import { TypedSelection } from "../typings/Types";
+import { performInsideOutsideAdjustment } from "./performInsideOutsideAdjustment";
+import { groupTargetsForEachEditor } from "./targetUtils";
 
 /** Unifies overlapping/intersecting ranges */
 export default function unifyRanges(ranges: Range[]): Range[] {
@@ -7,12 +10,12 @@ export default function unifyRanges(ranges: Range[]): Range[] {
   }
   let run = true;
   while (run) {
-    [ranges, run] = onePass(ranges);
+    [ranges, run] = unifyRangesOnePass(ranges);
   }
   return ranges;
 }
 
-function onePass(ranges: Range[]): [Range[], boolean] {
+function unifyRangesOnePass(ranges: Range[]): [Range[], boolean] {
   if (ranges.length < 2) {
     return [ranges, false];
   }
@@ -31,4 +34,81 @@ function onePass(ranges: Range[]): [Range[], boolean] {
     }
   });
   return [result, madeChanges];
+}
+
+/** Unifies overlapping/intersecting targets */
+export function unifyTargets(targets: TypedSelection[]): TypedSelection[] {
+  if (targets.length < 2) {
+    return targets;
+  }
+  return groupTargetsForEachEditor(targets).flatMap(([_editor, targets]) => {
+    if (targets.length < 2) {
+      return targets;
+    }
+    let results = [...targets];
+    results.sort((a, b) =>
+      a.selection.selection.start.compareTo(b.selection.selection.start)
+    );
+    let run = true;
+    // Merge targets untill there are no overlaps/intersections
+    while (run) {
+      [results, run] = unifyTargetsOnePass(results);
+    }
+    return results;
+  });
+}
+
+function unifyTargetsOnePass(
+  targets: TypedSelection[]
+): [TypedSelection[], boolean] {
+  if (targets.length < 2) {
+    return [targets, false];
+  }
+  const results: TypedSelection[] = [];
+  let currentGroup: TypedSelection[] = [];
+  targets.forEach((target) => {
+    // No intersection. Mark start of new group
+    if (
+      currentGroup.length &&
+      !intersects(currentGroup[currentGroup.length - 1], target)
+    ) {
+      results.push(mergeTargets(currentGroup));
+      currentGroup = [target];
+    } else {
+      currentGroup.push(target);
+    }
+  });
+  results.push(mergeTargets(currentGroup));
+  return [results, results.length !== targets.length];
+}
+
+function mergeTargets(targets: TypedSelection[]): TypedSelection {
+  if (targets.length === 1) {
+    return targets[0];
+  }
+  const first = targets[0];
+  const last = targets[targets.length - 1];
+  const typeSelection: TypedSelection = {
+    selection: {
+      editor: first.selection.editor,
+      selection: new Selection(
+        first.selection.selection.start,
+        last.selection.selection.end
+      ),
+    },
+    position: "contents",
+    selectionType: first.selectionType,
+    insideOutsideType: first.insideOutsideType,
+    selectionContext: {
+      leadingDelimiterRange: first.selectionContext.leadingDelimiterRange,
+      trailingDelimiterRange: last.selectionContext.trailingDelimiterRange,
+    },
+  };
+  return performInsideOutsideAdjustment(typeSelection);
+}
+
+function intersects(targetA: TypedSelection, targetB: TypedSelection) {
+  return !!targetA.selection.selection.intersection(
+    targetB.selection.selection
+  );
 }
